[REFACTOR] 의존성 주입, 엔티티 기본키, swagger 설명 추가 #24

+++ src/main/java/com/likelion/DSFest/repository/ImageRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ImageRepository extends JpaRepository<Image, Integer> {
-    public List<Image> findByNotice_NoticeId(Integer id);
-    public void deleteByNotice_NoticeId(Integer id);
+public interface ImageRepository extends JpaRepository<Image, Long> {
+    public List<Image> findByNotice_NoticeId(Long id);
+    public void deleteByNotice_NoticeId(Long id);
 }
